+++ pom.xml.versionsBackup
@@ -36,7 +36,7 @@
         <dependency>
             <groupId>org.apache.poi</groupId>
             <artifactId>poi</artifactId>
-            <version>4.1.2</version>
+            <version>5.2.2</version>
         </dependency>
         <dependency>
             <groupId>org.apache.commons</groupId>
