--- src/main/java/com/diceprojects/importcsv/config/WebConfig.java
@@ -1,11 +1,0 @@
-package com.diceprojects.importcsv.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@EnableWebMvc
-public class WebConfig implements WebMvcConfigurer {
-
-}
