--- src/main/java/com/diceprojects/importcsv/exceptions/CustomException.java
@@ -1,14 +1,0 @@
-package com.diceprojects.importcsv.exceptions;
-
-public class CustomException extends RuntimeException {
-    private final String errorMessage;
-
-    public CustomException(String errorMessage) {
-        super(errorMessage);
-        this.errorMessage = errorMessage;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-}
