--- src/main/java/com/diceprojects/importcsv/exceptions/ErrorResponse.java
@@ -1,17 +1,0 @@
-package com.diceprojects.importcsv.exceptions;
-
-public class ErrorResponse {
-    private String mensaje;
-
-    public ErrorResponse(String mensaje) {
-        this.mensaje = mensaje;
-    }
-
-    public String getMensaje() {
-        return mensaje;
-    }
-
-    public void setMensaje(String mensaje) {
-        this.mensaje = mensaje;
-    }
-}
