--- src/main/java/com/diceprojects/importcsv/exceptions/ToDoExceptions.java
@@ -1,21 +1,0 @@
-package com.diceprojects.importcsv.exceptions;
-
-import lombok.Data;
-import org.springframework.http.HttpStatus;
-
-
-@Data
-public class ToDoExceptions extends RuntimeException {
-
-    private String message;
-
-    private HttpStatus httpStatus;
-
-    public ToDoExceptions(String message, HttpStatus httpStatus) {
-
-        super(message);
-        this.message = message;
-        this.httpStatus = httpStatus;
-
-    }
-}
