+++ src/main/java/com/diceprojects/importfile/config/WebConfig.java
@@ -1,0 +1,14 @@
+package com.diceprojects.importfile.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+/**
+ * Configuración personalizada de Spring MVC.
+ */
+@Configuration
+@EnableWebMvc
+public class WebConfig implements WebMvcConfigurer {
+
+}
