+++ src/main/resources/application-dev.properties
@@ -1,5 +1,5 @@
 #Msvc configuration
-spring.application.name=msvc-import
+spring.application.name=msvc-importfile
 server.port=8002
 
 #BD Configuratiom
