--- src/test/java/com/diceprojects/importcsv/ImportCsvApplicationTests.java
@@ -1,13 +1,0 @@
-package com.diceprojects.importcsv;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ImportCsvApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
