+++ src/test/java/com/diceprojects/importfile/ImportFileApplicationTests.java
@@ -1,0 +1,13 @@
+package com.diceprojects.importfile;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ImportFileApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
